requirements.txt
requests
numpy
scikit-learn
faiss-cpu
pdfplumber
